order的异常处理

+++ src/main/java/com/homeward/webstore/service/order/OrderServiceImpl.java
@@ -1,24 +1,29 @@
 package com.homeward.webstore.service.order;
 
+import com.homeward.webstore.common.utils.CartUtil;
 import com.homeward.webstore.common.utils.JwtUtil;
+import com.homeward.webstore.mapper.AuthenticationMapper;
 import com.homeward.webstore.mapper.OrderMapper;
-import jakarta.servlet.http.Cookie;
+import com.homeward.webstore.mapper.StoreMapper;
 import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Map;
-import java.util.Set;
+import java.util.*;
 
 @Slf4j
 @Service
 public class OrderServiceImpl implements OrderService {
+    private final AuthenticationMapper authenticationMapper;
+    private final StoreMapper storeMapper;
     private final OrderMapper orderMapper;
 
 
-    public OrderServiceImpl(OrderMapper orderMapper) {
+    public OrderServiceImpl(AuthenticationMapper authenticationMapper, StoreMapper storeMapper, OrderMapper orderMapper) {
+        this.authenticationMapper = authenticationMapper;
+        this.storeMapper = storeMapper;
         this.orderMapper = orderMapper;
     }
 
@@ -37,8 +42,20 @@ public class OrderServiceImpl implements OrderService {
      * */
     @Override
     @Transactional
-    public void insertCart(Integer itemId) {
+    public void insertCart(Integer itemId, HttpServletRequest request) {
         String uuid = JwtUtil.getUserId();
+
+        if (storeMapper.getItemName(itemId) == null) {
+            throw new RuntimeException("item not found");
+        }
+
+        List<Integer> itemIdList = authenticationMapper.itemIdList(uuid, itemId);
+
+        CartUtil.isSingleColumn(itemIdList);
+
+        if (itemIdList.contains(itemId)) {
+            throw new RuntimeException("duplicated cart found");
+        }
 
         Integer itemAmount = 1;
 
@@ -50,24 +67,39 @@ public class OrderServiceImpl implements OrderService {
      * */
     @Override
     @Transactional
-    public void updateCart(Map<String, String> map) {
+    public void updateCart(Map<String, String> map, HttpServletRequest request) {
         String uuid = JwtUtil.getUserId();
 
-        String singleKey = "";
-        String amount = "";
-
-        Set<String> keySet = map.keySet();
+        Set<String> keySet =map.keySet();
+        String idString = null;
+        String value = null;
         for (String key : keySet) {
             if (key.contains("quantity")) {
-                singleKey = key;
-                amount = map.get(key);
+                value = map.get(key);
+                idString = key.replaceAll("[^0-9]", "");
                 break;
-            } else return;
+            }
         }
 
-        String orderId = singleKey.replaceAll("[^0-9]", "");
-        Integer itemId = Integer.parseInt(orderId);
-        Integer itemAmount = Integer.parseInt(amount);
+        if (StringUtils.isBlank(idString)) {
+            throw new RuntimeException("form data key not found");
+        }
+
+        if (Integer.parseInt(value) < 1 || Integer.parseInt(value) > 99) {
+            throw new RuntimeException("item amount out of bounds");
+        }
+
+        Integer itemAmount = Integer.parseInt(value);
+
+        Integer itemId = Integer.parseInt(idString);
+
+        List<Integer> itemIdList = authenticationMapper.itemIdList(uuid, itemId);
+
+        CartUtil.isSingleColumn(itemIdList);
+
+        if (!itemIdList.contains(itemId)) {
+            throw new RuntimeException("no such cart found");
+        }
 
         orderMapper.updateCart(uuid, itemId, itemAmount);
     }
@@ -79,8 +111,16 @@ public class OrderServiceImpl implements OrderService {
      */
     @Override
     @Transactional
-    public void deleteCart(Integer itemId) {
+    public void deleteCart(Integer itemId, HttpServletRequest request) {
         String uuid = JwtUtil.getUserId();
+
+        List<Integer> itemIdList = authenticationMapper.itemIdList(uuid, itemId);
+
+        CartUtil.isSingleColumn(itemIdList);
+
+        if (!itemIdList.contains(itemId)) {
+            throw new RuntimeException("no such cart found");
+        }
 
         orderMapper.deleteCart(uuid, itemId);
     }
